+++ cmake/cmake/Configuration.cmake
@@ -160,6 +160,15 @@ mark_as_advanced(PHP_LIBGCC)
 option(PHP_CCACHE "Use ccache if available on the system" ON)
 mark_as_advanced(PHP_CCACHE)
 
+cmake_dependent_option(
+  PHP_SYSTEM_GLOB
+  "Use the system glob() function instead of the PHP provided replacement"
+  OFF
+  [[NOT CMAKE_SYSTEM_NAME STREQUAL "Windows"]]
+  OFF
+)
+mark_as_advanced(PHP_SYSTEM_GLOB)
+
 ################################################################################
 # Set PHP_EXTENSION_DIR.
 ################################################################################